[feat]스탯 관련 private 접근을 위한 스크립트 수정
