+++ Assets/KSH/Scripts/PlayerStat/Fatigue.cs
@@ -1,19 +1,21 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using KSH;
 
 namespace KSH
 {
     public class Fatigue : MonoBehaviour
     {
         [SerializeField] private ResourceDegenerator resourceDegenerator;
-        
+
         public float FatigueCur => resourceDegenerator.Resource.Cur;
 
         public void Sleep(float amount)
         {
             resourceDegenerator.Resource.Increase(amount);
+        }
+
+        public void SetFatigue(float value)
+        {
+            resourceDegenerator.Resource.Cur = value;
         }
     }
 }
