+++ Assets/KSH/Scripts/PlayerStat/Thirsty.cs
@@ -1,20 +1,22 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using KSH;
 
 namespace KSH
 {
     public class Thirsty : MonoBehaviour
     {
         [SerializeField] private ResourceDegenerator resourceDegenerator;
-        
+
         public float ThirstyCur => resourceDegenerator.Resource.Cur;
 
 
         public void Drink(float amount)
         {
             resourceDegenerator.Resource.Increase(amount);
+        }
+
+        public void SetThirst(float value)
+        {
+            resourceDegenerator.Resource.Cur = value;
         }
     }
 }
